refactored node http server routes
